Trying to optimize and compress program size.
It's currently possible to compile up to 2 tasks (not any combination) at once. if all tasks are compiled, the program size is still ~3kB above the linker limit.

+++ utilities.c
@@ -73,8 +73,17 @@ void delay_us(unsigned int n) {
 	while(--o);
 }
 
+
+
+
+
+
+
+
+
 int int_to_char_1000(int input) {
-	uint8_t output = 0;
+	
+	int output = 0;
 	while(input >= 1000) {
 		input -= 1;
 		output += 1;
@@ -83,7 +92,7 @@ int int_to_char_1000(int input) {
 }
 
 int int_to_char_100(int input) {
-	uint8_t output = 0;
+	int output = 0;
 	while(input >= 100) {
 		input -= 1;
 		output += 1;
@@ -92,7 +101,7 @@ int int_to_char_100(int input) {
 }
 
 int int_to_char_10(int input) {
-	uint8_t output = 0;
+	int output = 0;
 	while(input >= 10) {
 		input -= 1;
 		output += 1;
@@ -101,7 +110,7 @@ int int_to_char_10(int input) {
 }
 
 int int_to_char_1(int input) {
-	uint8_t output = 0;
+	int output = 0;
 	while(input >= 1) {
 		input -= 1;
 		output += 1;
